Update README.md
I wanted to add that public health departments use the same data that Personicle is planning on using to make their stratification, yet their interventions are not the most effective since each individual will only find a proportion of the total solution valid as departments have to develop population-based solutions that are off the greatest benefit to the greatest number of people. This does not leave room for nuance hence the interventions are not as effective for an individual as they are for a population. Since healthcare is trending towards an individualized approach, I thought it best to describe that what we are doing is not invasive as local governments have been doing it for many years.